gzip/python/pyproject.toml
[project]
name = "gzip-sandbox"
version = "0.1.0"
description = "Default template for PDM package"
authors = [{ name = "Uy Ha", email = "hchanuy@gmail.com" }]
dependencies = [
    "pyzmq>=25.1.2",
    "pyserde[msgpack]>=0.14.0",
]
requires-python = "==3.12.*"
readme = "README.md"
license = { text = "MIT" }


[tool.pdm]
distribution = false
